budget: React app shell + entry + LegacyRoute bridge (Unit 2 of #1874)
Replace the imperative main.ts mount with a React root: main.tsx createRoot,
App.tsx composing ds Nav + header + footer + content-grid + a minimal client
router (use-router.ts), AuthControls in Nav's end slot, a Hero island, and a
LegacyRoute bridge that string-renders + hydrates every current route exactly as
before. The data/crypto/file-sync/FSA pipeline moves verbatim into
use-app-state.ts; only its DOM glue becomes React state. Per-route hydration uses
the preserved error-wrapper (legacy-hydrate.ts). LegacyRoute keys by route+navEpoch
so same-path data transitions re-resolve the body, matching the legacy
router.navigate() behavior. main.test.ts is ported to use-app-state.test.tsx,
preserving every FSA/plaintext-downgrade/external-reload/startup-branch assertion.
