+++ budget/src/main.tsx
@@ -1,0 +1,9 @@
+import "missing.css";
+import "./style/theme.css";
+import { createRoot } from "react-dom/client";
+import { App } from "./App.js";
+
+// createRoot (not hydrateRoot): the body is a single <div id="root"> in
+// index.html; the React tree is mounted client-side. (Unit 6 reworks the
+// prerender step; ds token CSS imports land in Unit 5.)
+createRoot(document.getElementById("root")!).render(<App />);
